added a second more complex fit example
The example is taken from an email we recieved a few weeks ago.
In this example a sigmoid function has to be fitted to the data.
I find this to be a nice example for a fit that that does not work
if no initial values are supplied.
In addition I added some cells to explain a possible way of 'guessing'
the right parameters.