reduced number of non f-string examples
With formatting done in the f-string section there is not
any need for too much examples for format(). I reduced
it to two examples showing the basic usage and the usage of
named parameters.
I left the latex formatting example in this section because
the string can either an f-string or an r-string.