added example for evaluation of a fake-experiment
This is kind of a guided tour on how to use the different
features to your advantage. The features do not get added
at once but step by step in different versions of the same
evaluation script.

It starts out by using Python as a calculator
and then uses more and more Python features (lists, dicts, loops)
in a more or less realistic to see how these features
can be used in the 'real world'.

The files 'evaluation-i.py' contain the script in its iterations
and also explanations what happens from step to step.
The files 'evaluation-i_Hinweise.py' contain some hints for what
to do next.